requirements.txt
# Core LangChain / LangGraph
langchain>=0.3.0
langgraph>=0.2.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.4

# Vector store
chromadb>=0.5.0

# BM25 file-level triage
rank-bm25>=0.2.2

# PDF extraction + OCR
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Embeddings
sentence-transformers>=3.0.0

# Excel I/O
openpyxl>=3.1.0
pandas>=2.0.0

# Config
python-dotenv>=1.0.0

# Validation
pydantic>=2.0.0
